requirements.txt
aws-cdk-lib==2.59.0
constructs>=10.0.0,<11.0.0
pytest==6.2.5
aws_cdk.aws_glue_alpha
## test cases
boto3
aws-lambda-powertools
moto
mockito
boto3
pytz
pandas
fastparquet
mock
s3fs